py disassembler

+++ src/me/enerccio/sp/utils/Coerce.java
@@ -163,7 +163,7 @@ public class Coerce {
 			@Override
 			public Object coerce(PythonObject o, Class<?> clazz) throws CastFailedException {
 				if (o instanceof NumberObject)
-					return Integer.valueOf(((NumberObject)o).getJavaInt().intValue());
+					return Integer.valueOf((int)((NumberObject)o).getJavaInt());
 				
 				throw new CastFailedException("Can't convert " + o.toString() + " to Integer");
 			}
@@ -174,7 +174,7 @@ public class Coerce {
 			@Override
 			public Object coerce(PythonObject o, Class<?> clazz) throws CastFailedException {
 				if (o instanceof NumberObject)
-					return ((NumberObject)o).getJavaInt().intValue();
+					return (int)((NumberObject)o).getJavaInt();
 				
 				throw new CastFailedException("Can't convert " + o.toString() + " to int");
 			}
@@ -197,7 +197,7 @@ public class Coerce {
 			@Override
 			public Object coerce(PythonObject o, Class<?> clazz) throws CastFailedException {
 				if (o instanceof NumberObject)
-					return ((NumberObject)o).getJavaInt().longValue();
+					return ((NumberObject)o).getJavaInt();
 				
 				throw new CastFailedException("Can't convert " + o.toString() + " to long");
 			}
